requirements.txt
Django==3.1.7
django-mptt
